refactor: MVC 패턴 구사 위해 패키지화. feat:로또 번호 비교 기능, enum활용
TODO: 요구사항 라이브러리 사용해야 함.예외 구사애햐 함.
