+++ docs/README.md
@@ -2,6 +2,6 @@ Application : 메인 실행 클래스
 
 LottoService : 로또 생성 및 계산
 
-LottoResult(ENUM 사용) : 당첨 결과 관리
+LottoRankENUM 사용) : 당첨 결과 관리
 
 Lotto: 사용자 번호 저장
