+++ src/test/java/lottoView/ApplicationTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lottoView;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
