+++ src/test/java/lottoView/LottoTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lottoView;
 
+import lottoModel.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
